hihiep

+++ admin/category/list.php
@@ -1,8 +1,8 @@
 <h3 class="alert alert-secondary">Thêm loại món</h3>
-<form action="index.php" method="post" id="add_category" class="form m-b-2">
+<form action="index.php" method="post" id="form_insert_category" class="form m-b-2">
     <div class="form-group">
         <label for="">Tên loại hàng</label>
-        <input type="text" name="category_name" class="form-control" aria-describedby="helpId">
+        <input type="text" name="category_name" id='category_name' class="form-control" aria-describedby="helpId">
     </div>
     <div class="form-group">
         <button type="submit" name="btn_insert" class="btn btn-dark">Thêm </button>
@@ -32,7 +32,7 @@
                             <td><?= $category['category_name'] ?></td>
                             <td>
                                 <a href="index.php?btn_edit&category_id=<?= $category['category_id'] ?>" class="btn btn-defaule">Sửa</a>
-                                <a href="index.php?btn_delete&category_id=<?= $category['category_id'] ?>" class="btn btn-defaule">Xóa</a>
+                                <a href="index.php?btn_delete&category_id=<?= $category['category_id'] ?>" onclick=" return confirm('Bạn có chắc chắn muốn xóa không ???')" class="btn btn-defaule">Xóa</a>
                             </td>
                         </tr>
                     <?php endforeach ?>
